Use size_t for the public threaded scheduler
Change the public scheduler callback indices and task count to `size_t`,
and preserve that type through `parallel_batches`. This aligns scheduler
task ranges with the `size_t` dimensions and stripe ranges used across
KleidiCV, avoiding narrowing before work is scheduled. Remove the
explicit `unsigned` narrowing from add-padding batching.

Update the fake and example schedulers for the new signatures. Since
OpenCV `parallel_for_` uses `int` ranges, use it for task counts up to
`INT_MAX` and fall back to a serial callback for larger ranges.

Add regression coverage verifying that `parallel_batches` passes a
dimension greater than `UINT_MAX` to the scheduler without narrowing it.
Update the threading documentation and changelog for the public API
change.

This changes the public threading scheduler source and ABI contract.

+++ kleidicv_thread/include/kleidicv_thread/kleidicv_thread.h
@@ -52,8 +52,8 @@ extern "C" {
 /// @param end Index one past the last task in the range, exclusive.
 /// @param data Opaque pointer supplied by KleidiCV. The scheduler must pass it
 ///             unchanged to this callback.
-typedef kleidicv_error_t (*kleidicv_thread_callback)(unsigned begin,
-                                                     unsigned end, void *data);
+typedef kleidicv_error_t (*kleidicv_thread_callback)(size_t begin, size_t end,
+                                                     void *data);
 
 /// Signature of a function to invoke callbacks in parallel.
 ///
@@ -81,7 +81,7 @@ typedef kleidicv_error_t (*kleidicv_thread_callback)(unsigned begin,
 /// @param task_count Number of task indices requested by KleidiCV.
 typedef kleidicv_error_t (*kleidicv_thread_parallel)(
     kleidicv_thread_callback callback, void *callback_data, void *parallel_data,
-    unsigned task_count);
+    size_t task_count);
 
 /// Encapsulates the caller-provided threading backend.
 typedef struct {
